:recycle: Timer관련 로직 useTimer훅으로 추상화 및 resetTimer함수 내용 수정(stopTimer활용)
